requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.22.0
pillow>=9.0.0
tqdm>=4.65.0
matplotlib>=3.5.0
pyyaml>=6.0
scikit-learn>=1.2.0
safetensors>=0.3.0

# Diffusion model
denoising-diffusion-pytorch>=1.0.0

# Optional but recommended
wandb>=0.15.0  # For experiment tracking
tensorboard>=2.11.0  # Alternative logging
